1st attempt at serial focuser class
got as far as connect()
