+++ src/panoptes/pocs/focuser/birger.py
@@ -5,7 +5,7 @@ import glob
 from warnings import warn
 from contextlib import suppress
 
-from panoptes.pocs.focuser import AbstractFocuser
+from panoptes.pocs.serial import AbstractSeriaFocuser
 from panoptes.utils import error
 
 # Birger adaptor serial numbers should be 5 digits
@@ -42,7 +42,7 @@ error_messages = ('No error',
                   'Distance stops not supported by the lens')
 
 
-class Focuser(AbstractFocuser):
+class Focuser(AbstractSerialFocuser):
     """
     Focuser class for control of a Canon DSLR lens via a Birger Engineering Canon EF-232 adapter.
 
@@ -63,18 +63,15 @@ class Focuser(AbstractFocuser):
     # Class variable to cache the device node scanning results
     _birger_nodes = None
 
-    # Class variable to store the device nodes already in use. Prevents scanning known Birgers &
-    # acts as a check against Birgers assigned to incorrect ports.
-    _assigned_nodes = []
-
     def __init__(self,
                  name='Birger Focuser',
                  model='Canon EF-232',
                  initial_position=None,
                  dev_node_pattern='/dev/tty.USA49*.?',
                  *args, **kwargs):
-        super().__init__(name=name, model=model, *args, **kwargs)
         self.logger.debug('Initialising Birger focuser')
+
+        self.baudrate = 115200
 
         if serial_number_pattern.match(self.port):
             # Have been given a serial number
@@ -117,26 +114,8 @@ class Focuser(AbstractFocuser):
             self.logger.debug('Found {} ({}) on {}'.format(self.name, self.port, device_node))
             self.port = device_node
 
-        # Check that this node hasn't already been assigned to another Birgers
-        if self.port in Focuser._assigned_nodes:
-            message = 'Device node {} already in use!'.format(self.port)
-            self.logger.error(message)
-            warn(message)
-            return
+        super().__init__(name=name, model=model, *args, **kwargs)
 
-        try:
-            self.connect(self.port)
-        except (serial.SerialException,
-                serial.SerialTimeoutException,
-                AssertionError) as err:
-            message = 'Error connecting to {} on {}: {}'.format(self.name, self.port, err)
-            self.logger.error(message)
-            warn(message)
-            return
-
-        Focuser._assigned_nodes.append(self.port)
-        self._is_moving = False
-        self._initialise()
         if initial_position is not None:
             self.position = initial_position
 
@@ -152,16 +131,6 @@ class Focuser(AbstractFocuser):
     ##################################################################################################
     # Properties
     ##################################################################################################
-
-    @property
-    def is_connected(self):
-        """
-        Checks status of serial port to determine if connected.
-        """
-        connected = False
-        if self._serial_port:
-            connected = self._serial_port.isOpen()
-        return connected
 
     @AbstractFocuser.position.getter
     def position(self):
@@ -216,36 +185,8 @@ class Focuser(AbstractFocuser):
     ##################################################################################################
 
     def connect(self, port):
-        try:
-            # Configure serial port.
-            # Settings copied from Bob Abraham's birger.c
-            self._serial_port = serial.Serial()
-            self._serial_port.port = port
-            self._serial_port.baudrate = 115200
-            self._serial_port.bytesize = serial.EIGHTBITS
-            self._serial_port.parity = serial.PARITY_NONE
-            self._serial_port.stopbits = serial.STOPBITS_ONE
-            self._serial_port.timeout = 2.0
-            self._serial_port.xonxoff = False
-            self._serial_port.rtscts = False
-            self._serial_port.dsrdtr = False
-            self._serial_port.write_timeout = None
-            self._inter_byte_timeout = None
 
-            # Establish connection
-            self._serial_port.open()
-
-        except serial.SerialException as err:
-            self._serial_port = None
-            self.logger.critical('Could not open {}!'.format(port))
-            raise err
-
-        # Want to use a io.TextWrapper in order to have a readline() method with universal newlines
-        # (Birger sends '\r', not '\n'). The line_buffering option causes an automatic flush() when
-        # a write contains a newline character.
-        self._serial_io = io.TextIOWrapper(io.BufferedRWPair(self._serial_port, self._serial_port),
-                                           newline='\r', encoding='ascii', line_buffering=True)
-        self.logger.debug('Established serial connection to {} on {}.'.format(self.name, port))
+        self._connect()
 
         # Set 'verbose' and 'legacy' response modes. The response from this depends on
         # what the current mode is... but after a power cycle it should be 'rm1,0', 'OK'
